Refactor search page model
