+++ tests/01-end-to-end.sh
@@ -8,9 +8,9 @@ TESTCAFE_TESTS_FOLDER="$(pwd)/end-to-end"
 startMigrationAssetsContainer
 
 # 'Admin' tests
-docker run --rm --network gateway --env-file "${ENV_FILE}" -v "${TESTCAFE_TESTS_FOLDER}":/tests testcafe/testcafe --screenshots path=/tests/screenshots,takeOnFails=true chromium /tests/tests/admin/**/*.spec.js
+docker run --rm --network gateway --env-file "${ENV_FILE}" -v "${TESTCAFE_TESTS_FOLDER}":/tests testcafe/testcafe --screenshots path=/tests/reports/screenshots,takeOnFails=true chromium /tests/tests/admin/**/*.spec.js
 # Add UI test data
-docker run --rm --network gateway --env-file "${ENV_FILE}" -v "${TESTCAFE_TESTS_FOLDER}":/tests testcafe/testcafe --screenshots path=/tests/screenshots,takeOnFails=true chromium /tests/tests/ui/data-migrations.js
-docker run --rm --network gateway --env-file "${ENV_FILE}" -v "${TESTCAFE_TESTS_FOLDER}":/tests testcafe/testcafe --screenshots path=/tests/screenshots,takeOnFails=true chromium /tests/tests/ui/data-media.js
+docker run --rm --network gateway --env-file "${ENV_FILE}" -v "${TESTCAFE_TESTS_FOLDER}":/tests testcafe/testcafe --screenshots path=/tests/reports/screenshots,takeOnFails=true chromium /tests/tests/ui/data-migrations.js
+docker run --rm --network gateway --env-file "${ENV_FILE}" -v "${TESTCAFE_TESTS_FOLDER}":/tests testcafe/testcafe --screenshots path=/tests/reports/screenshots,takeOnFails=true chromium /tests/tests/ui/data-media.js
 # 'UI' tests
-docker run --rm --network gateway --env-file "${ENV_FILE}" -v "${TESTCAFE_TESTS_FOLDER}":/tests testcafe/testcafe --screenshots path=/tests/screenshots,takeOnFails=true chromium /tests/tests/ui/*.spec.js
+docker run --rm --network gateway --env-file "${ENV_FILE}" -v "${TESTCAFE_TESTS_FOLDER}":/tests testcafe/testcafe --screenshots path=/tests/reports/screenshots,takeOnFails=true chromium /tests/tests/ui/*.spec.js
